fix(bpf): widen disk_event.nr_bytes to __u64 to prevent overflow
ctx->nr_sector * 512 was computed in __u32, wrapping for any merged or
discard request larger than ~8 MiB (nr_sector > 8_388_607). Cast the
operand to __u64 so the multiply is done in 64-bit arithmetic.

Widen the nr_bytes field in struct disk_event (kerno.h) from __u32 to
__u64 and re-pad _pad from [3]byte to [7]byte to keep the struct size a
multiple of 8. Mirror the new layout in the Go DiskEvent struct and
update the bpf2go-generated field order accordingly.

Add TestDecodeDiskEventLargeIO to catch regressions: a 16 MiB discard
(nr_sector=32768) previously decoded to NrBytes=0.

Fixes: disk_io.c:55 __u32 overflow on large merged/discard requests

+++ internal/bpf/c/disk_io.c
@@ -52,7 +52,7 @@ int tracepoint_block_rq_complete(struct trace_event_raw_block_rq_completion *ctx
     e->latency_ns   = latency;
     e->sector        = sector;
     e->dev           = (__u32)ctx->dev;
-    e->nr_bytes      = ctx->nr_sector * 512;
+    e->nr_bytes      = (__u64)ctx->nr_sector * 512;
     e->pid           = (__u32)(bpf_get_current_pid_tgid() >> 32);
 
     // rwbs[0] is the primary op (R/W/D); subsequent positions hold flag
